refactor: organize remaining root components into feature directories
