+++ src/components/analysis/index.ts
@@ -1,0 +1,3 @@
+export { PatchAnalysis } from './patch-analysis';
+export { VulnerabilityScatterplot } from './vulnerability-scatterplot';
+export { SectionCards } from './section-cards';
