+++ src/components/analysis/patch-analysis.tsx
