+++ src/components/analysis/section-cards.tsx
