+++ src/components/analysis/vulnerability-scatterplot.tsx
