+++ src/components/images/DockerImageAutocomplete.tsx
