+++ src/components/images/docker-image-selector.tsx
