+++ src/components/scan/findings/VulnerabilitiesTab.tsx
@@ -26,7 +26,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import { VulnerabilityUrlMenu } from "@/components/vulnerability-url-menu";
+import { VulnerabilityUrlMenu } from "@/components/shared/vulnerability-url-menu";
 import { getSeverityCssClass } from "@/lib/utils/severity-utils";
 
 interface VulnerabilitiesTabProps {
